refactor: 투표 리스트 응답 DTO를 기능별 패키지로 분리 (active, mine, participated) (#37)
